ImageClassification/web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
tensorflow
keras
requests
numpy
pillow